ai-api/requirements.txt
flask==3.0.0
flask-cors==4.0.0
langchain==0.3.12
langchain-openai==0.2.14
langchain-google-genai==2.0.8
langchain-community==0.3.12
langchain-core==0.3.28
pydantic==2.10.5
python-dotenv==1.0.0
youtube-transcript-api==0.6.2
pytube==15.0.0